+++ epoxy-composesample/src/main/java/com/airbnb/epoxy/compose/sample/MainActivity.kt
@@ -10,11 +10,12 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.airbnb.epoxy.EpoxyRecyclerView
-import com.airbnb.epoxy.compose.composableInterop
+import com.airbnb.epoxy.composableInterop
 import com.airbnb.epoxy.compose.sample.epoxyviews.headerView
 
 class MainActivity : ComponentActivity() {
@@ -23,6 +24,7 @@ class MainActivity : ComponentActivity() {
         setContentView(R.layout.activity_main)
 
         val recyclerView: EpoxyRecyclerView = findViewById(R.id.interop_recycler_view)
+        val composeView: ComposeView = findViewById(R.id.compose_view)
 
         recyclerView.withModels {
             headerView {
@@ -30,21 +32,32 @@ class MainActivity : ComponentActivity() {
                 title("Testing Composable in Epoxy")
             }
 
-            composableInterop(
-                id = "compose_text",
-                composeFunction = {
-                    ShowCaption(
-                        modifier =
-                        Modifier.background(color = Color.Blue)
-                            .padding(16.dp)
-                    )
-                }
-            )
+            for (i in 0..100) {
+                composableInterop(
+                    id = "compose_text_$i",
+                    composeFunction = {
+                        ShowCaption(
+                            modifier =
+                            Modifier.background(color = Color.Blue)
+                                .padding(16.dp)
+                        )
+                    }
+                )
 
-            composableInterop(
-                id = "news",
-                composeFunction = { NewsStory() }
-            )
+                composableInterop(
+                    id = "news_$i",
+                    composeFunction = { NewsStory() }
+                )
+            }
+
+
+        }
+
+        composeView.setContent {
+            Text("Testing epoxy in Composable")
+//            headerView {
+//                setTitle("title")
+//            }
         }
     }
 }
